+++ pycognaize/document/field/__init__.py
@@ -3,6 +3,7 @@ import enum
 from .field import Field
 from .date_field import DateField
 from .numeric_field import NumericField
+from .section_field import SectionField
 from .table_field import TableField
 from .text_field import TextField
 from .area_field import AreaField
@@ -19,3 +20,4 @@ class FieldMapping(enum.Enum):
     text = TextField
     table = TableField
     area = AreaField
+    section = SectionField