+++ pycognaize/document/snapshot.py
@@ -1,6 +1,8 @@
 import os
 from typing import Mapping
 
+from pycognaize.common import utils
+from pycognaize.common.exceptions import AuthenthicationError
 from pycognaize.common.enums import EnvConfigEnum
 from pycognaize.common.lazy_dict import LazyDocumentDict
 from pycognaize.login import Login
@@ -21,10 +23,24 @@ class Snapshot:
         return self._documents
 
     @classmethod
-    def _snapshot_path(cls, remote_snapshot_root: str = None) -> str:
+    def download(cls, snapshot_id: str, destination_dir: str):
+        """Downloads snapshot to specified destination"""
+        login_instance = Login()
+
+        if login_instance.logged_in:
+            snapshot_path = os.path.join(login_instance.snapshot_root,
+                                         snapshot_id)
+            ci = utils.cloud_interface_login(login_instance)
+            ci.copy_dir(snapshot_path, destination_dir)
+        else:
+            raise AuthenthicationError()
+
+    @classmethod
+    def _snapshot_path(cls) -> str:
         """Identify and return the snapshot path"""
-        if remote_snapshot_root:
-            snapshot_dir = remote_snapshot_root
+        login_instance = Login()
+        if login_instance.logged_in:
+            snapshot_dir = login_instance.snapshot_root
             snapshot_id = os.environ[EnvConfigEnum.SNAPSHOT_ID.value]
             snapshot_path = os.path.join(snapshot_dir, snapshot_id)
         else:
@@ -36,11 +52,4 @@ class Snapshot:
     @classmethod
     def get(cls) -> 'Snapshot':
         """Read the snapshot object from local storage and return it"""
-        login_instance = Login()
-        if login_instance.logged_in:
-            remote_snapshot_root = login_instance.snapshot_root
-            return cls(path=cls._snapshot_path(
-                       remote_snapshot_root=remote_snapshot_root))
-
-        else:
-            return cls(path=cls._snapshot_path())
+        return cls(path=cls._snapshot_path())